Rename Tags file. Fix tag values loading in dropdown
